requirements.txt
numpy==2.3.4
alpaca-py==0.43.2